requirements.txt
requests~=2.31.0
psycopg2-binary~=2.9.10
schedule~=1.2.2
flask~=3.1.0
python-dateutil==2.9.0